WIKI-1107: [Regression] Space group instead of Space name in Welcome message
